extend CanBeAnt computation
Rationale: According to the SSAPRE paper, "`can_be_avail` is false for a phi if
and only if no down-safe placement of computations can make the expression
available."

In this particular case, it's possible for an unsplittable critical edge to
prevent filling a null def. Filling can't happen regardless of up-safety, so
`CanBeAnt` should be false by vacuous truth: No up-safe placement can make the
store anticipated because no store can be safely placed.