fix naive critical edge splitting bug
it's possible for more than one edge to exist from a lambda block to a given
successor. in such cases, the edges need to be transferred to the new block
generated by critical edge splitting. this is exactly what `MergeIdenticalEdges`
does.

furthermore, the critical edge detection function, "hasMultiplePreds," isn't
enough because for our purposes, the edge head block needs to have multiple
_distinct_ predecessors in order to be critical.