bug: loop-variant red classes should die above ptr def
`%r` could potentially take on a different value each time the bb2 loop is
exited, which means that `%loc` could have different values on each iteration of
bb3. So the transformation is unsafe.

The problem here is that PDSE only leaves top-of-block kills at parent blocks of
phis that are themselves SCC members. The only such phi here is in bb2. But in
fact, `%loc` should count as dead above bb3.