use debug counters to control when PRE is performed
Usage scenario: Suppose we've encountered a mis-compile with pdse. Bugpoint and
srcbisect.py will pinpoint the problem to a single function. However, the sheer
size of the function itself may still prevent manual analysis.

Consider the set of all partial redundancy transformations that occur within the
function. We know that one or more from this set causes the mis-compile. With
debug counters, we can selectively disable a portion of these transformation
sites and narrow down to the exact problem site.

The format of the command lline invocation is:

```
opt -pdse \
    --debug-counter=pdse-partial-elim-skip=2000,pdse-partial-elim-count=0 \
    source-file.ll
```